requirements.txt
# Core Libraries
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
SQLAlchemy==2.0.28
requests==2.31.0
tqdm==4.66.2

# For Data Analysis & Notebooks
jupyter==1.0.0
matplotlib==3.8.3
seaborn==0.13.2

# For Web Interface (optional, can be added later)
# flask==3.0.2
# jinja2==3.1.3